[Refactor] move A2C in separate folder for clarity
